+++ src/main/java/duke/parser/Parser.java
@@ -1,16 +1,16 @@
 package duke.parser;
 
-import duke.commands.ByeCommand;
-import duke.commands.Command;
-import duke.commands.DeadlineCommand;
-import duke.commands.DeleteCommand;
-import duke.commands.DoneCommand;
-import duke.commands.EventCommand;
-import duke.commands.ListCommand;
-import duke.commands.TodoCommand;
+import duke.command.ByeCommand;
+import duke.command.Command;
+import duke.command.DeadlineCommand;
+import duke.command.DeleteCommand;
+import duke.command.DoneCommand;
+import duke.command.EventCommand;
+import duke.command.ListCommand;
+import duke.command.TodoCommand;
 import duke.enums.CommandEnum;
-import duke.exceptions.DukeException;
-import duke.tasks.Task;
+import duke.exception.DukeException;
+import duke.task.Task;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
